requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
fastapi==0.115.0
uvicorn==0.32.0
fastmcp==2.13.0
pandas==2.3.1
numpy==1.26.4
openpyxl==3.1.5
xlrd==2.0.1
xlwt==1.3.0
xlutils==2.0.0
pytest==8.3.4
mysql-connector-python==9.0.0
oracledb==3.4.0
vanna==0.7.9
requests==2.32.4
sentence-transformers==3.0.1
transformers==4.44.2
chromadb==0.5.4
matplotlib==3.9.2
torch==2.4.1
langchain==0.2.5
langchain-community==0.2.5
langchain-text-splitters==0.2.0
langchain-openai==0.1.7
jieba==0.42.1
FlagEmbedding==1.2.11
google-genai==1.10.0
